tools/requirements.apt.txt
python3-intelhex
python3-prettytable
python3-future
python3-jinja2
python3-click
python3-git
python3-tqdm
python3-tabulate
python3-requests
python3-psutil
python3-pyudev
python3-typing-extensions
python3-serial
python3-dotenv
python3-appdirs
python3-fasteners
python3-lockfile
python3-junit.xml
python3-cryptography
python3-cbor
